Fixed for Linux

+++ Source/Auravyx/Audio/WAVE.cpp
@@ -31,18 +31,18 @@ void WAVE::load(const char * path)
 	fp = (FILE*)fopen(path, "rb");
 	
 	char type[4] = { 0, 0, 0, 0 };
-	DWORD size, chunkSize;
+	unsigned int size, chunkSize;
 	short formatType, channels;
-	DWORD sampleRate, avgBytesPerSec;
+    unsigned int sampleRate, avgBytesPerSec;
 	short bytesPerSample, bitsPerSample;
-	DWORD dataSize;
+    unsigned int dataSize;
 
 	fread(type, sizeof(char), 4, fp);
 	if (type[0] != 'R' || type[1] != 'I' || type[2] != 'F' || type[3] != 'F')
 	{
 		Log::error("[WAVE Loader] No 'RIFF' found!");
 	}
-	fread(&size, sizeof(DWORD), 1, fp);
+	fread(&size, sizeof(unsigned int), 1, fp);
 	fread(type, sizeof(char), 4, fp);
 	
 	if (type[0] != 'W' || type[1] != 'A' || type[2] != 'V' || type[3] != 'E')
@@ -54,11 +54,11 @@ void WAVE::load(const char * path)
 	{
 		Log::error("[WAVE Loader] No 'fmt' found!");
 	}
-	fread(&chunkSize, sizeof(DWORD), 1, fp);
+	fread(&chunkSize, sizeof(unsigned int), 1, fp);
 	fread(&formatType, sizeof(short), 1, fp);
 	fread(&channels, sizeof(short), 1, fp);
-	fread(&sampleRate, sizeof(DWORD), 1, fp);
-	fread(&avgBytesPerSec, sizeof(DWORD), 1, fp);
+	fread(&sampleRate, sizeof(unsigned int), 1, fp);
+	fread(&avgBytesPerSec, sizeof(unsigned int), 1, fp);
 	fread(&bytesPerSample, sizeof(short), 1, fp);
 	fread(&bitsPerSample, sizeof(short), 1, fp);
 
@@ -68,10 +68,10 @@ void WAVE::load(const char * path)
 		Log::out("OpenAL", "Audio device missing 'data' (" + std::to_string(type[0]) + "" + std::to_string(type[1])
 			+ "" + std::to_string(type[2]) + "" + std::to_string(type[3]) + ")", RED);
 	}
-	fread(&dataSize, sizeof(DWORD), 1, fp);
+	fread(&dataSize, sizeof(unsigned int), 1, fp);
 
 	unsigned char* buf = new unsigned char[dataSize];
-	fread(buf, sizeof(BYTE), dataSize, fp);
+	fread(buf, sizeof(unsigned int), dataSize, fp);
 
 	if (bitsPerSample == 8)
 	{
